refactor(geode-core): group linear/iteration machinery under solver [4/10]
Make the linear-solve and iteration machinery directory-backed under a
new `solver` module (epic #377, child 4/10; mirrors merged PR #390):

- `git mv ksp_solve.rs -> solver/ksp.rs` (`solver::ksp`)
- `git mv iterate.rs -> solver/iterate.rs` (`solver::iterate`)
- New `solver/mod.rs` with group-level `//!` rustdoc.

lib.rs: drop `pub mod ksp_solve;` / `pub mod iterate;`, add `pub mod
solver;`; replace the two plain `pub use` blocks with
`#[deprecated]` flat-root item shims (all 16 public items) sourced
from the canonical `solver::ksp` / `solver::iterate` paths.

prelude.rs: add high-traffic solver entries from canonical paths.

Same-PR caller migration (all lib-target `src/`): `crate::ksp_solve::*`
-> `crate::solver::ksp::*` in driven.rs/extraction.rs; `crate::iterate`
-> `crate::solver::iterate` imports + intra-doc links in
mie.rs/silvermuller_self_consistent.rs/lanczos.rs/complex_lanczos.rs.

Pure refactor — no behavior change. File history preserved via git mv.

Closes #381
