+++ crates/geode-core/src/lanczos.rs
@@ -27,12 +27,12 @@
 //! Convergence is declared when the residual norm
 //! `‖K x - λ M x‖_2 / ‖λ M x‖_2 < tol` for **every** requested mode.
 //!
-//! # Why this loop is not ported onto [`crate::iterate`]
+//! # Why this loop is not ported onto [`crate::solver::iterate`]
 //!
 //! Full reorthogonalization keeps the **entire** basis history, so the
 //! Krylov basis `V_k` gains one column per Lanczos iteration — the
 //! carried state grows by one vector each step. That violates
-//! [`crate::iterate`] **contract restriction 1** (loop-invariant
+//! [`crate::solver::iterate`] **contract restriction 1** (loop-invariant
 //! carried-state shapes): a trace-once graph backend traces the loop body
 //! once and cannot express a state slot whose tensor shape changes
 //! between iterations. This is exactly the "do not grow a `Vec` inside
