+++ crates/geode-core/src/prelude.rs
@@ -2,7 +2,7 @@
 //!
 //! Re-exports from the canonical module paths (`crate::assembly`,
 //! `crate::backend`, `crate::derham`, `crate::elements`,
-//! `crate::traits`) — never the deprecated root shims — so
+//! `crate::solver`, `crate::traits`) — never the deprecated root shims — so
 //! glob-importing the prelude stays warning-free under `-D warnings`.
 //! Later children of the namespace reorg add their own groups here.
 pub use crate::assembly::fe::{DirichletBc, ElementType, FeAssembleResult, fe_assemble};
@@ -15,4 +15,6 @@ pub use crate::elements::nedelec::{
     NedelecLocalMatrices, batched_nedelec_local_matrices, tet_edges,
 };
 pub use crate::elements::p1::{P1LocalMatrices, batched_p1_local_matrices};
+pub use crate::solver::iterate::{IterOutcome, Step, iterate_while, iterate_while_with_prev};
+pub use crate::solver::ksp::{Cocg, JacobiPreconditioner, KspReport, KspSolve};
 pub use crate::traits::{Element, Mesh, Operator};
