+++ crates/geode-core/src/solver/iterate.rs
