+++ crates/geode-core/src/solver/ksp.rs
